fix bugs, change enemy stats
there's smth about the gameplay loop that feels stale that the reload definitely didn't fix. idk if it's bc the atks are all very similar, or if there's not any opportunity cost to using them or what. I'm thinking of merging the atk and special decks and giving the player 2 cards at a time or smth. Maybe throw in a few dash cards in the base deck to make movement better